# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Escrituras",
        "description": "API para acceder a las escrituras de La Iglesia de Jesucristo de los Santos de los Últimos Días",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/libros/{id}": {
            "get": {
                "tags": [
                    "Libros"
                ],
                "summary": "Obtiene un libro específico",
                "operationId": "7ba0f98c8ba5780dd795ab4081054868",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del libro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Libro encontrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Libro"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Libro no encontrado"
                    }
                }
            }
        },
        "/api/libros/{id}/partes": {
            "get": {
                "tags": [
                    "Libros"
                ],
                "summary": "Lista todas las partes de un libro",
                "operationId": "d12157c1f59fd708214bc642abffa222",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del libro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de partes obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Parte"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Libro no encontrado"
                    }
                }
            }
        },
        "/api/libros/{id}/capitulos": {
            "get": {
                "tags": [
                    "Libros"
                ],
                "summary": "Lista todos los capítulos de un libro",
                "operationId": "a5ce5e0c4addeec6f1513a885b710c8f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del libro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de capítulos obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Capitulo"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Libro no encontrado"
                    }
                }
            }
        },
        "/api/libros/{id}/division": {
            "get": {
                "tags": [
                    "Libros"
                ],
                "summary": "Obtiene la división a la que pertenece el libro",
                "operationId": "0f04263035a84b9c3e0b53982c28e563",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del libro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "División obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Division"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Libro no encontrado"
                    }
                }
            }
        },
        "/api/libros/{id}/volumen": {
            "get": {
                "tags": [
                    "Libros"
                ],
                "summary": "Obtiene el volumen al que pertenece el libro",
                "operationId": "f219e43bef7ffeb043dba3b57b58b87f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del libro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Volumen obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Volumen"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Libro no encontrado"
                    }
                }
            }
        },
...
        "/api/partes/{id}": {
            "get": {
                "tags": [
                    "Partes"
                ],
                "summary": "Obtiene una parte específica",
                "operationId": "2d194b8b14fbc3440017a75e2682cf85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la parte",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Parte encontrada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Parte"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Parte no encontrada"
                    }
                }
            }
        },
        "/api/partes/{id}/capitulos": {
            "get": {
                "tags": [
                    "Partes"
                ],
                "summary": "Lista todos los capítulos de una parte",
                "operationId": "10760a1872bed2571d3ce7337ce8be59",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la parte",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de capítulos obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Capitulo"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Parte no encontrada"
                    }
                }
            }
        },
        "/api/partes/{id}/libro": {
            "get": {
                "tags": [
                    "Partes"
                ],
                "summary": "Obtiene el libro al que pertenece la parte",
                "operationId": "d4bae71d22dccf48cb63aa6ed70278db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la parte",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Libro obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Libro"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Parte no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Volumen": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "abreviatura": {
                        "type": "string"
                    },
                    "orden": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Division": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "abreviatura": {
                        "type": "string"
                    },
                    "orden": {
                        "type": "integer"
                    },
                    "volumen_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Libro": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "abreviatura": {
                        "type": "string"
                    },
                    "orden": {
                        "type": "integer"
                    },
                    "division_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Parte": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "orden": {
                        "type": "integer"
                    },
                    "libro_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Capitulo": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "num_capitulo": {
                        "type": "integer"
                    },
                    "orden": {
                        "type": "integer"
                    },
                    "libro_id": {
                        "type": "integer"
                    },
                    "parte_id": {
                        "type": "integer",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}